+++ Presentation.Web/Controllers/External/V2/ItSystemUsages/ItSystemUsageRelationFrequencyTypeV2Controller.cs
@@ -12,10 +12,10 @@ using Swashbuckle.Swagger.Annotations;
 
 namespace Presentation.Web.Controllers.External.V2.ItSystemUsages
 {
-    [RoutePrefix("api/v2/it-system-usage-relation-frequencies")]
-    public class ItSystemUsageRelationFrequenciesController : BaseOptionTypeV2Controller<SystemRelation, RelationFrequencyType>
+    [RoutePrefix("api/v2/it-system-usage-relation-frequency-types")]
+    public class ItSystemUsageRelationFrequencyTypeV2Controller : BaseOptionTypeV2Controller<SystemRelation, RelationFrequencyType>
     {
-        public ItSystemUsageRelationFrequenciesController(IOptionsApplicationService<SystemRelation, RelationFrequencyType> optionService)
+        public ItSystemUsageRelationFrequencyTypeV2Controller(IOptionsApplicationService<SystemRelation, RelationFrequencyType> optionService)
             : base(optionService)
         {
         }
