+++ Presentation.Web/Presentation.Web.csproj
@@ -349,7 +349,7 @@
     <Compile Include="Controllers\External\V2\ItSystemUsages\ItSystemUsageArchiveTypeV2Controller.cs" />
     <Compile Include="Controllers\External\V2\ItSystemUsages\ItSystemUsageDataClassificationTypeV2Controller.cs" />
     <Compile Include="Controllers\External\V2\ItSystems\ItSystemV2Controller.cs" />
-    <Compile Include="Controllers\External\V2\ItSystemUsages\ItSystemUsageRelationFrequenciesController.cs" />
+    <Compile Include="Controllers\External\V2\ItSystemUsages\ItSystemUsageRelationFrequencyTypeV2Controller.cs" />
     <Compile Include="Models\Users\UserWithCrossOrganizationalRightsDTO.cs" />
     <Compile Include="Models\Users\UserWithOrganizationDTO.cs" />
     <Compile Include="Extensions\DtoModelV2MappingExtensions.cs" />
