+++ Tests.Integration.Presentation.Web/Tools/External/OptionV2ApiHelper.cs
@@ -14,8 +14,8 @@ namespace Tests.Integration.Presentation.Web.Tools.External
         public static class ResourceName
         {
             public const string BusinessType = "business-types";
-            public const string ItSystemUsageDataClassification = "it-system-usage-data-classifications";
-            public const string ItSystemUsageRelationFrequencies = "it-system-usage-relation-frequencies";
+            public const string ItSystemUsageDataClassification = "it-system-usage-data-classification-types";
+            public const string ItSystemUsageRelationFrequencies = "it-system-usage-relation-frequency-types";
             public const string ItSystemUsageArchiveTypes = "it-system-usage-archive-types";
         }
